cleanup of base services strategy. Both backend use the same abstract base strategy class
