+++ otelib/backends/python/base.py
@@ -5,13 +5,15 @@ from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING
 from uuid import uuid4
 
+from otelib.backends.strategies import AbstractBaseStrategy
 from otelib.exceptions import ApiError
-from otelib.pipe import Pipe
 from otelib.settings import Settings
 
 if TYPE_CHECKING:  # pragma: no cover
     from pathlib import Path
     from typing import Optional
+
+    from otelib.pipe import Pipe
 
 
 class Singleton(object):
@@ -27,7 +29,7 @@ class Cache(Singleton, dict):
     pass
 
 
-class BasePythonStrategy(ABC):
+class BasePythonStrategy(AbstractBaseStrategy):
     """Abstract class for strategies for the python backend.
 
     Parameters:
@@ -63,65 +65,6 @@ class BasePythonStrategy(ABC):
         # For debugging/testing
         self.debug: bool = bool(os.getenv("OTELIB_DEBUG", ""))
         self._session_id: "Optional[str]" = None
-
-    @abstractmethod
-    def create(self, **kwargs) -> None:
-        """Create a strategy.
-
-        It should post the configuration for the created strategy.
-        """
-
-    @abstractmethod
-    def fetch(self, session_id: str) -> bytes:
-        """Returns the result of the current strategy.
-
-        This method is called by `get()` while propagating up the pipeline.
-
-        Parameters:
-            session_id: The ID of the session shared by the pipeline.
-
-        Returns:
-            The response from the OTEAPI Service.
-
-        """
-
-    @abstractmethod
-    def initialize(self, session_id: str) -> bytes:
-        """Initialise the current strategy.
-
-        This method is called by `get()` when propagating down the pipeline.
-
-        Parameters:
-            session_id: The ID of the session shared by the pipeline.
-
-        Returns:
-            The response from the OTEAPI Service.
-
-        """
-
-    def _set_input(self, input_pipe: Pipe) -> None:
-        """Used by `__rshift__` to set the input pipe.
-
-        Parameters:
-            input_pipe: A pipe representing the strategy that is "piped" into this
-                strategy.
-
-        """
-        self.input_pipe = input_pipe
-
-    def __rshift__(self, other: "BasePythonStrategy") -> "BasePythonStrategy":
-        """Implements strategy concatenation using the `>>` symbol.
-
-        Parameters:
-            other: The next strategy this one is "piping" into.
-
-        Returns:
-            The next strategy this one is "piped" into.
-
-        """
-        pipe = Pipe(self)
-        other._set_input(pipe)
-        return other
 
     def get(self, session_id: "Optional[str]" = None) -> bytes:
         """Executes a pipeline.
